Added TD Sarsa
I also made the code look a bit cleaner and put the code to run a certain model into its own function so we just have to comment one line each time instead of three, just QOL stuff. Also there's a .spyproject in there now because I need it for Spyder but it doesn't change anything. I'm going to be heavily modifying the reward system now so I wanna push this so I can rollback if I need to.